implement hook for `mkinitcpio`
The files  `{hooks,install}/untouchable` need to be installed inside the
`/etc/initcpio/{hooks,install}/`, or in case installation is done with a
package manager, the `/usr/lib/initcpio/{hooks,install}/` folders.

This will enable  `mkinitcpio` to include the hook `untouchable`,  if so
configured via  `/etc/mkinitcpio.conf` and the per-kernel presets inside
`/etc/mkinitcpio.d/`.